+++ src/ExpeditionsMacro.App/Pages/IAppPage.cs
@@ -4,5 +4,5 @@ public interface IAppPage
 {
     Task OnShownAsync();
 
-    Func<Task>? IdleF6Action { get; }
+    Func<Task>? IdleHotkeyAction { get; }
 }